JSP게시판 : 구현 완료

+++ WebContent/view.jsp
@@ -9,6 +9,7 @@
 <meta charset="UTF-8 ">
 <meta name="viewport" content="width=device-width", initial-scale="1">
 <link rel="stylesheet" href="css/bootstrap.css">
+<link rel="stylesheet" href="css/custom.css">
 <title>JSP 게시판 웹사이트</title>
 </head>
 <body>
@@ -113,7 +114,7 @@
 					if(userID != null && userID.equals(bbs.getUserID())){
 				%>
 					<a href="update.jsp?bbsID=<%=bbsID %>" class="btn btn-primary">수정</a>
-					<a href="deleteAction.jsp?bbsID=<%=bbsID %>" class="btn btn-primary">삭제</a>
+					<a onclick="return confirm('삭제하시겠습니까?')" href="deleteAction.jsp?bbsID=<%=bbsID %>" class="btn btn-primary">삭제</a>
 				<%					
 					}
 				%>
